Refactored, added rand(), updated README
* Refacted files to suit new types and functions, removing unused code.
* Added rand(::Rotation), which samples 3D rotations uniformly using the
  Haar measure (i.e. the rotationally-invariant distribution).
* The README has been changed to reflect the new approach.
